[0027] Prevent sending transport maps in player update object (will be replaced by a better fix later). Should fix the random teleports for this moment...
Fixed aura update (Fixes warrior/paladin client freeze on world login).

Conflicts:
	src/game/Map.cpp
	src/game/Transports.cpp

+++ src/game/SpellAuras.cpp
@@ -10937,7 +10937,7 @@ void SpellAuraHolder::BuildUpdatePacket(WorldPacket& data) const
     data << uint32(GetId());
 
     uint8 auraFlags = GetAuraFlags();
-    data << uint8(auraFlags);
+    data << uint16(auraFlags);
     data << uint8(GetAuraLevel());
     data << uint8(GetStackAmount() > 1 ? GetStackAmount() : (GetAuraCharges()) ? GetAuraCharges() : 1);
